requirements.txt
gunicorn==20.0.4
astroid==2.4.2
certifi==2020.11.8
click==7.1.2
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
easyocr==1.2.1
fastapi==0.61.2
future==0.18.2
h11==0.11.0
imageio==2.9.0
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pydantic==1.7.2
pylint==2.6.0
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.1
python-multipart==0.0.5
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
starlette==0.13.6
tifffile==2020.11.18
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
typing-extensions==3.7.4.3
uvicorn==0.12.3
wrapt==1.11.2
